# Optimize the python client performance for Python 3.10 or later