--- a/iotdb-client/client-py/iotdb/utils/NumpyTablet.py
+++ b/iotdb-client/client-py/iotdb/utils/NumpyTablet.py
@@ -102,7 +102,8 @@ class NumpyTablet(object):
         bs_len = 0
         bs_list = []
         for i, value in enumerate(self.__values):
-            if self.__data_types[i] == TSDataType.TEXT:
+            # TEXT
+            if self.__data_types[i].value == 5:
                 format_str_list = [">"]
                 values_tobe_packed = []
                 for str_list in value:
@@ -118,6 +119,7 @@ class NumpyTablet(object):
                     values_tobe_packed.append(value_bytes)
                 format_str = "".join(format_str_list)
                 bs = struct.pack(format_str, *values_tobe_packed)
+            # Non-TEXT
             else:
                 bs = value.tobytes()
             bs_list.append(bs)