--- a/iotdb-client/client-py/iotdb/utils/Tablet.py
+++ b/iotdb-client/client-py/iotdb/utils/Tablet.py
@@ -18,7 +18,6 @@
 
 import struct
 
-from iotdb.utils.IoTDBConstants import TSDataType
 from iotdb.utils.BitMap import BitMap
 
 
@@ -95,7 +94,7 @@ class Tablet(object):
         for i in range(self.__column_number):
             bitmap = None
             bitmaps.append(bitmap)
-            data_type_value = int(self.__data_types[i])
+            data_type_value = self.__data_types[i]
             if data_type_value == 0:
                 format_str_list.append(str(self.__row_number))
                 format_str_list.append("?")