## Description

TBinaryProtocolAccelerated is a C-Accelerated version of TBinaryProtocol. Using it instead of TBinaryProtocol can improve the Thrift transport performance. 

_Notice: The test on my mac shows that the acceleration works on Python 3.10 and 3.11. Python 3.7, 3.8 or 3.9 **cannot** accelerate!_

Besides, IntEnum has better performance than normal Enum, we need to use it to implement TSDataType, TSEncoding, Compressor.

## How to test

Run the following code with Python 3.10 or 3.11.
```python
from datetime import datetime
import numpy as np

from iotdb.Session import Session
from iotdb.utils.IoTDBConstants import TSDataType, TSEncoding, Compressor
from iotdb.utils.NumpyTablet import NumpyTablet

# creating session connection.
ip = "127.0.0.1"
port_ = "6667"
username_ = "root"
password_ = "root"
session = Session(ip, port_, username_, password_, fetch_size=1024, zone_id="UTC+8", enable_redirection=False)

session.open(False)

device = "root.sg.d"
measurements_ = []
ts_path_lst_ = []
values_ = []
for i in range(440):
    measurement = "s" + str(i)
    measurements_.append(measurement)
    ts_path_lst_.append(device + "." + measurement)
    values_.append(1.1 * i)

data_type_lst_ = [TSDataType.FLOAT for _ in range(440)]
encoding_lst_ = [TSEncoding.GORILLA for _ in range(440)]
compressor_lst_ = [Compressor.LZ4 for _ in range(440)]
startTime = int(datetime.now().timestamp() * 1000)
for i in range(10000):
    session.test_insert_record(device, 1692761800000 + i, measurements_, data_type_lst_, values_)
print("InsertRecord 10000 rows cost: " + str(int(datetime.now().timestamp() * 1000) - startTime) + "ms")
exit(0)
```
## Test result
### Python 3.10
Before this PR: InsertRecord 10000 rows cost: **5163ms**

After: InsertRecord 10000 rows cost: **1451ms**

### Python 3.11
Before this PR: InsertRecord 10000 rows cost: **5058ms**

After: InsertRecord 10000 rows cost: **1237ms**